src/video_insert/pyproject.toml
[tool.poetry]
name = "video-insert"
version = "0.1.0"
description = "Сервис video_insert проекта vidos. Отвечает за загрузку видео на сервис."
authors = ["Andreev Aisen  <isenandreev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.12"
email-validator = "1.3.1"
fastapi = "0.95.0"
minio = "7.1.14"
motor = "3.1.2"
passlib = "1.7.4"
python-jose = "3.3.0"
pytz = "2023.3"
requests = "2.28.2"
setuptools = "67.6.1"
uvicorn = "0.21.1"
wheel = "0.40.0"
python-multipart = "0.0.6"
python-dotenv = "^1.0.0"
bcrypt = "^4.1.2"
pymongo = "^4.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
